+++ PixelPusher-objc/Source/PPPixel.h
@@ -6,21 +6,21 @@
 //  Copyright (c) 2013 rrrus. All rights reserved.
 //
 
-#import <Foundation/Foundation.h>
+#import <UIKit/UIKit.h>
 
 @class UIColor;
 
 @interface PPPixel : NSObject
 
-@property (nonatomic, assign) float red;
-@property (nonatomic, assign) float green;
-@property (nonatomic, assign) float blue;
+@property (nonatomic, assign) CGFloat red;
+@property (nonatomic, assign) CGFloat green;
+@property (nonatomic, assign) CGFloat blue;
 
-+ (PPPixel*)pixelWithHue:(float)hue saturation:(float)sat luminance:(float)lum;
-+ (PPPixel*)pixelWithRed:(float)r green:(float)g blue:(float)b;
++ (PPPixel*)pixelWithHue:(CGFloat)hue saturation:(CGFloat)sat luminance:(CGFloat)lum;
++ (PPPixel*)pixelWithRed:(CGFloat)r green:(CGFloat)g blue:(CGFloat)b;
 
-- (PPPixel*)pixelScalingLuminance:(float)scale;
-- (void)scaleLuminance:(float)scale;
+- (PPPixel*)pixelScalingLuminance:(CGFloat)scale;
+- (void)scaleLuminance:(CGFloat)scale;
 - (void)addPixel:(PPPixel*)other;
 - (void)copy:(PPPixel*)other;
 - (void)setColor:(UIColor*)color;
